[B-834] Remove attachments from other work panel (#846)
